+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+### Changed
+- [Patch] Pass `isActive` prop on `TabNav.Tab` instead of doing the logic on the child. (#522)
 
 ## [15.0.0][15.0.0] - 2016-08-26
 ### Changed